customize to my liking
Use Alpine as a base.
Remove nginx and the related functionality.
Bootstrap Akaunting install into a writable volume on first run.
Use volume mounted sockets rather than tcp for communicating with nginx.
Simplify and reorganize a bunch of the scripts.
Store configurable values in a .env file and ignore with .gitignore.